Create delivery_bot.c
In this project, a robot uses AVR-Atmega328p Microcontroller (Arduino) to make an Auto-delivery 
process using line detection principles with IR sensors. The robot loads goods at the initial point, and 
by following the line using IR sensors, the robot reaches the destination point, unload the goods, and
returns to the initial point for subsequent delivery.